Update README.md
I almost began to describe one of the pictures and look up the plants in it. I suppose it behooves me to voice expectations for further commits. I do expect a branch. I am fond of the style of the automatically created. Not only do I also like foreshadowing in commit messages. Not only did I create this repository and add a read me. I also downloaded the existing tutorial POC that show routing to www.oldmill.click. But I also appreciate descriptions of context. Because it makes the narrative of events in perspective. I obviously going to next upload to s3 a changed html file. I may have to figure out the routing what didn't require index.html.