Add files via upload
The links are hypothetical to near coded in HTML. We still need pages for the photo contest. Perhaps, a forgotten memories section to house found photos. Lucy's from the beyond, Crafts and Decor. We should get the color working. Yes index.htm. Upload these files to s3 and that version control. Synch up when trivially noteworthy be why not. Don't forget to call tungstendiadem a colorful fellow and an adequate developer.